+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 ## 3.2.17
 - Add `price` in `AtsReceptionProduct`
+- Add `atsDriver`, `atsTerminalOperations`, `atsTerminalLoadingGrid` , `atsTerminalAccessControl`, `atsTerminalYardAlerts`, `atsTerminalLoadFormatting`,
+`atsTerminalStockManagement`, `atsTerminalCacl`,
+`atsTerminalEditTransport`, `atsAccessColeta`,
+`atsPaLoadComboio` in `ReferenceCategory`
+
 
 ## 3.2.16